Extract QuoteTypeTags
